MISC: linter

+++ main.py
@@ -33,14 +33,16 @@ async def on_private_ban(message: types.Message):
         period = 360
     user = int(user)
     period = float(period)
-    until = math.floor(time.time()) + period * 60 * 60
-    await bot.restrict_chat_member(chat_id=-1001338616632,
-                                   user_id=user,
-                                   until_date=until,
-                                   can_send_messages=False,
-                                   can_send_media_messages=False,
-                                   can_send_other_messages=False,
-                                   can_add_web_page_previews=False)
+    until = int(math.floor(time.time()) + period * 60 * 60)
+    await bot.restrict_chat_member(
+        chat_id=-1001338616632,
+        user_id=user,
+        until_date=until,
+        can_send_messages=False,
+        can_send_media_messages=False,
+        can_send_other_messages=False,
+        can_add_web_page_previews=False
+    )
 
 
 @dp.message_handler(lambda msg: is_private_command(msg, 'register'))
